netdev-rte-offloads: tun meta data changed to id to save meta data in HW.
The assumption is that the number of tun entities is rather small up to
few thousadns. We can replace each connection (3-tuple), with an
outer_id, and when we have a match on flow tun, we can replace it with a
match with the id.
When flow with tunnel pop is offloaded we should set the id in the right
register.